Removed `ServerListener` and migrated logic to `WorldListener`
Consolidated event handling by moving overworld load logic into `WorldListener` for better maintainability. Adjusted `levelView` type in `WorldsPlugin` accordingly.
